Several Changes
1. Changed the R in contact point
2. In one step optimization, remove the constraint on y final velocity
(doesn't have to be 0) and y average velocity(doen't have to be 0). With
the confident that the two step y velocity will still be zero finally.
3. Use default ipopt_options for now. (gaits are much more 'optimal'
than the modified options) May change in the future.
4. Manually remove default discretemap in impact for safe.
5. For symFunction, add tomatrix() to all expression to make sure the
size will be correct. (Sometimes if the size is not correct the
constraint may not be added, I don't know why for now)